refactor: move signup to components
