+++ webapp/src/containers/SideBar.js
@@ -17,7 +17,7 @@ import { useTranslation } from 'react-i18next'
 import LoginModal from '../components/LoginModal'
 import CredentialsRecovery from '../components/CredentialsRecovery'
 import TokenTransfer from '../components/TokenTransfer'
-import Signup from '../routes/Signup/Signup'
+import Signup from '../components/Signup/Signup'
 
 const useStyles = makeStyles((theme) => ({
   wrapper: {
